backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.5
qdrant-client==1.9.1
python-dotenv==1.0.0
PyPDF2==3.0.1
tiktoken==0.7.0
langchain==0.1.16
langchain-community==0.0.38
langchain-openai==0.1.3
SQLAlchemy==2.0.23
pydantic==2.5.0
gradio==4.28.3